Limit ElasticBeanstalk to 'Ready' Environments
Address the [issue][1] wherein 'Terminated' Environments are being reported
as Classic EC2 resources that need to be addressed.  Which is not true.
This commit filters on those resources that are 'Ready' so that if they
have Classic EC2 resources they will be properly reported.

By submitting this pull request, I confirm that you can use, modify, copy, and
redistribute this contribution, under the terms of your choice.

[1]:  https://github.com/aws-samples/ec2-classic-resource-finder/issues/3